63-CleanCode/src/test/java/chapter16/refactoring/MonthTest.java
package chapter16.refactoring;

import junit.framework.TestCase;


public class MonthTest extends TestCase {
    public void testMonthToQuarter() throws Exception {
        assertEquals(1, Month.JANUARY.quarter());
        assertEquals(1, Month.FEBRUARY.quarter());
        assertEquals(1, Month.MARCH.quarter());
        assertEquals(2, Month.APRIL.quarter());
        assertEquals(2, Month.MAY.quarter());
        assertEquals(2, Month.JUNE.quarter());
        assertEquals(3, Month.JULY.quarter());
        assertEquals(3, Month.AUGUST.quarter());
        assertEquals(3, Month.SEPTEMBER.quarter());
        assertEquals(4, Month.OCTOBER.quarter());
        assertEquals(4, Month.NOVEMBER.quarter());
        assertEquals(4, Month.DECEMBER.quarter());

        try {
            Month.fromInt(-1);
            fail("Invalid Month code should throw exception");
        } catch (IllegalArgumentException e) {
        }
    }
}